requirements.txt
-i https://pypi.org/simple
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
cron-descriptor==1.3.0
cycler==0.11.0
django==4.2.1
django-admin-soft-dashboard==1.0.12
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-debug-toolbar==4.0.0
django-timezone-field==5.0
djangorestframework==3.14.0
fonttools==4.39.4
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
matplotlib==3.7.1
numpy==1.23.5
packaging==23.1
pandas==2.0.1
pillow==9.5.0
prompt-toolkit==3.0.38
psycopg[binary]==3.1.9
psycopg-binary==3.1.9
pyaml==23.5.8
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pyyaml==6.0
redis==4.5.5
scikit-learn==1.2.2
scikit-optimize==0.9.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
typing-extensions==4.5.0 ; python_version >= '3.7'
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
xgboost==1.7.5
